+++ lib/ecs_deploy_cli/cli.rb
@@ -1,11 +1,24 @@
+# frozen_string_literal: true
+
 module EcsDeployCli
   class CLI < Thor
+    def self.exit_on_failure?
+      true
+    end
+
     desc 'validate', 'Validates your ECSFile'
     option :file, default: 'ECSFile'
     def validate
       @parser = load(options[:file])
       runner.validate!
       puts 'Your ECSFile looks fine! 🎉'
+    end
+
+    desc 'diff', 'Check differences between task definitions'
+    option :file, default: 'ECSFile'
+    def diff
+      @parser = load(options[:file])
+      runner.diff
     end
 
     desc 'version', 'Updates all services defined in your ECSFile'
@@ -29,13 +42,28 @@ module EcsDeployCli
       runner.update_services! timeout: options[:timeout], service: options[:only]
     end
 
-    desc 'deploy', 'Updates a single service defined in your ECSFile'
+    desc 'deploy', 'Updates all services and scheduled tasks at once'
     option :file, default: 'ECSFile'
     option :timeout, type: :numeric, default: 500
     def deploy
       @parser = load(options[:file])
       runner.update_services! timeout: options[:timeout]
       runner.update_crons!
+    end
+
+    desc 'run-task NAME', 'Manually runs a task defined in your ECSFile'
+    option :launch_type, default: 'FARGATE'
+    option :security_groups, default: '', type: :string
+    option :subnets, required: true, type: :string
+    option :file, default: 'ECSFile'
+    def run_task(task_name)
+      @parser = load(options[:file])
+      runner.run_task!(
+        task_name,
+        launch_type: options[:launch_type],
+        security_groups: options[:security_groups].split(','),
+        subnets: options[:subnets].split(',')
+      )
     end
 
     desc 'ssh', 'Connects to ECS instance via SSH'