+++ lib/ecs_deploy_cli/dsl/auto_options.rb
@@ -1,9 +1,25 @@
+# frozen_string_literal: true
+
 module EcsDeployCli
   module DSL
     module AutoOptions
+      extend ActiveSupport::Concern
+
+      module ClassMethods
+        def allowed_options(*value)
+          @allowed_options = value
+        end
+
+        def _allowed_options
+          @allowed_options ||= []
+        end
+      end
+
       def method_missing(name, *args, &block)
         if args.count == 1 && !block
-          EcsDeployCli.logger.info("Auto-added option security_group #{name.to_sym} = #{args.first}")
+          unless self.class._allowed_options.include?(name)
+            EcsDeployCli.logger.info("Used unhandled option #{name.to_sym} = #{args.first} in #{self.class.name}")
+          end
           _options[name.to_sym] = args.first
         else
           super