+++ CHANGELOG.md
@@ -1,7 +1,15 @@
 # Changelog
 All notable changes to this project made by Monade Team are documented in this file. For info refer to team@monade.io
 
-## [0.1.0-alpha] - 2021-03-21
+## [0.2.0] - 2021-03-25
+### Added
+- Command run-task
+- Command diff
+- Made automatic options explicit
+- Deploy scheduled tasks
+- Refactoring runners
+
+## [0.1.0] - 2021-03-21
 First functional version
 
 ### Added